Refactor job submission machinery
